+++ packages/Input/src/index.d.ts
@@ -5,8 +5,8 @@ interface InputProps {
   [x: string]: any;
   /** Provides a non-visible label for this input for assistive technologies. */
   a11yText?: string;
-  /** Sets the class for the input. */
-  className?: string;
+  /** Optional Input.Container to collect props for root DOM element. */
+  children?: React.ReactNode;
   /** Custom icon for the clear action in the input. */
   clearIcon?: React.ReactNode;
   /** Sets the default input value for an uncontrolled component. */
@@ -21,17 +21,13 @@ interface InputProps {
   isDisabled?: boolean;
   /** If true it makes the input read only. */
   isReadOnly?: boolean;
-  /** Callback to be executed when the input value is changed. Receives the
-onChange event as an argument, except when the clear button is clicked,
-then the argument is null. Needed when value prop is provided (component
-is controlled). */
+  /** Callback to be executed when the input value is changed. Receives the onChange event as an argument, except when the clear button is clicked, then the argument is null. Required when value prop is provided (component is controlled). */
   onChange?: (...args: any[]) => any;
   /** Changes the size of the input. */
   size?: Input.types.size.SMALL | Input.types.size.MEDIUM | Input.types.size.LARGE;
   /** Allows user to specify the type of input. */
   type?:
     | Input.types.type.EMAIL
-    | Input.types.type.NUMBER
     | Input.types.type.PASSWORD
     | Input.types.type.SEARCH
     | Input.types.type.TELEPHONE
@@ -54,7 +50,6 @@ declare namespace Input {
   namespace types {
     namespace type {
       const EMAIL: any;
-      const NUMBER: any;
       const PASSWORD: any;
       const SEARCH: any;
       const TELEPHONE: any;