+++ packages/Input/stories/Input.stories.js
@@ -1,43 +1,39 @@
 import React from "react";
-import { storiesOf } from "@storybook/react";
 import { withKnobs } from "@storybook/addon-knobs";
 import { getStoryName } from "storybook/storyTree";
-import ExampleStory from "storybook/components/ExampleStory";
+import { showcaseStoryParameters, variationsStoryParameters } from "storybook/assets/storyParameters";
 import { InputStory } from "./Input.stories.styles";
-import InputExample from "./examples/InputExample";
-import ShowcaseStory from "./examples/Showcase";
-import SizesStory from "./examples/Sizes";
-import WithContentStory from "./examples/WithContent";
-import WithIconStory from "./examples/WithIcon";
-import WithCustomClearIconStory from "./examples/WithCustomClearIconStory";
-import WithDisabledReadOnlyStory from "./examples/WithDisabledReadOnly";
-import TypesStory from "./examples/Types";
-import WithRef from "./examples/WithRef";
-import Uncontrolled from "./examples/Uncontrolled";
-import Input from "../src/Input";
+import Showcase from "./examples/Showcase";
+import Variations from "./examples/Variations";
 
 const storyName = getStoryName("Input");
 
-storiesOf(storyName, module)
-  .addDecorator(withKnobs)
-  .add("Showcase", () => (
-    <ExampleStory storyName="Input" tagline={ExampleStory.defaultTaglines.showcase}>
-      <ShowcaseStory />
-    </ExampleStory>
-  ));
+export default {
+  title: storyName,
+};
 
-storiesOf(`${storyName}/Examples`, module)
-  .add("Sizes", () => <SizesStory />)
-  .add("With content", () => <WithContentStory />)
-  .add("With decorative icon", () => <WithIconStory />)
-  .add("With custom clear icon", () => <WithCustomClearIconStory />)
-  .add("With isDisabled / isReadOnly", () => <WithDisabledReadOnlyStory />)
-  .add("Types", () => <TypesStory />)
-  .add("With Ref", () => <WithRef />)
-  .add("Uncontrolled", () => <Uncontrolled />);
-
-storiesOf(`${storyName}/Backyard/Tests`, module).add("Screener", () => (
-  <InputStory>
-    <InputExample placeholder="First Name" size={Input.types.size.LARGE} />
+export const showcase = () => (
+  <InputStory storyName="Input" tagline={InputStory.defaultTaglines.showcase}>
+    <Showcase />
   </InputStory>
-));
+);
+showcase.story = {
+  name: "Showcase",
+  decorators: [withKnobs],
+  parameters: showcaseStoryParameters,
+};
+
+export const variations = () => (
+  <InputStory
+    storyName="Input"
+    tagline={InputStory.defaultTaglines.variations}
+    component="Input"
+    fileName="examples/Variations.js"
+  >
+    <Variations />
+  </InputStory>
+);
+variations.story = {
+  name: "Variations",
+  parameters: variationsStoryParameters,
+};