+++ packages/Panel/src/index.d.ts
@@ -90,7 +90,7 @@ declare namespace Panel {
 
     children?: (...args: any[]) => any;
     /** container of the Overlay element */
-    container?: instanceOf;
+    container?: React.ReactNode;
 
     focusLockOptions?: shape;