+++ packages/ActionBar/CHANGELOG.md
@@ -1,5 +1,47 @@
 # Changelog
 
+## 3.1.1-next.0
+
+### Patch Changes
+
+- Updated dependencies [45d341f]
+- Updated dependencies [14bd5f5]
+  - @paprika/input@4.0.0-next.0
+  - @paprika/popover@1.2.2-next.0
+
+## 3.1.0
+
+### Minor Changes
+
+- e6b7618: Support saving order in localStorage for ColumnsArrangement
+
+### Patch Changes
+
+- ab296ef: Support moreProps for ColumnsArrangement
+- Updated dependencies [5d0db59]
+- Updated dependencies [9ad9c79]
+  - @paprika/icon@1.1.0
+  - @paprika/helpers@2.1.2
+  - @paprika/button@1.0.9
+  - @paprika/input@3.0.1
+  - @paprika/popover@1.2.1
+  - @paprika/select@1.0.7
+  - @paprika/sortable@1.1.8
+  - @paprika/switch@1.0.8
+
+## 3.1.0-next.3
+
+### Patch Changes
+
+- Updated dependencies [9ad9c79]
+  - @paprika/helpers@2.1.2-next.0
+  - @paprika/button@1.0.9-next.1
+  - @paprika/input@3.0.1-next.1
+  - @paprika/popover@1.2.1-next.0
+  - @paprika/select@1.0.7-next.0
+  - @paprika/sortable@1.1.8-next.1
+  - @paprika/switch@1.0.8-next.0
+
 ## 3.1.0-next.2
 
 ### Minor Changes