+++ packages/ExternalLink/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 2.0.6
+
+### Patch Changes
+
+- Updated dependencies [5d0db59]
+  - @paprika/icon@1.1.0
+
 ## 2.0.6-next.0
 
 ### Patch Changes