+++ packages/Tabs/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 2.2.2
+
+### Patch Changes
+
+- Updated dependencies [9ad9c79]
+  - @paprika/helpers@2.1.2
+
+## 2.2.2-next.0
+
+### Patch Changes
+
+- Updated dependencies [9ad9c79]
+  - @paprika/helpers@2.1.2-next.0
+
 ## 2.2.1
 
 ### Patch Changes